Added GoogleTest testing framework
Added the gtest (but not gmock) files and
include headers from googletest, and wrote a simple
CMakeLists.txt to build it, so you do not have to download
and install it separately. We have also added CMake macros
so you can add tests in subdirectories - all of them will
be executed if you run 'make check'.