Changed organization and project name details
After forking the entire GitHub repository from
the eriklindahl to the IHPCSS git account and changing
the name, we also need to change a bunch of internal
names to reflect that. This commit also serves as an
example of the change you would need to make after
cloning the repo to make your own program.

We will still need to set up the Travis and
ReadTheDocs builds manually, so for now the builds will
not pass.